notification, awareness, blog, imp doc

+++ controller/admin/awaremessController.js
@@ -1,9 +1,23 @@
 const AwarenessCategory = require("../../model/awareness/awarenessCategoryModel");
 const Awareness = require("../../model/awareness/awarenessModel");
+const { UploadImage } = require("../../utils/imageUpload");
+const { deleteImgFromFolder } = require("../../utils/removeImages");
 
 const createCategory = async (req, res) => {
+  // console.log("================== createCategory =================");
+  // console.log("req.body: ", req.body);
+  // console.log("req.files: ", req.files);
+  const name = req.body?.name
+  const image = req.files?.image
+
+
   try {
-    const response = new AwarenessCategory(req.body);
+    const response = new AwarenessCategory({ name });
+    if (image) {
+      const fileName = await UploadImage(image, "awarenessCategory");
+      response.image = fileName;
+    }
+
     const saveresponse = await response.save();
     res.status(201).json({ success: true, data: saveresponse, message: "Awareness category Created", });
   } catch (error) {
@@ -12,7 +26,22 @@ const createCategory = async (req, res) => {
 };
 
 const UpdateCategory = async (req, res) => {
+  // const name = req.body?.name
+  const image = req.files?.image
+
   try {
+    const checkAwarenessCategory = await AwarenessCategory.findById(req.params.id)
+    if (!checkAwarenessCategory) {
+      return res.status(404).json({ success: false, message: "Category not found" });
+    }
+    if (image) {
+      // const oldImage = checkAwarenessCategory.image;
+      const fileName = await UploadImage(image, "awarenessCategory");
+      req.body.image = fileName;
+      if (checkAwarenessCategory.image) {
+        await deleteImgFromFolder(checkAwarenessCategory.image, "awarenessCategory");
+      }
+    }
     const response = await AwarenessCategory.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true, });
     if (response) {
       return res.status(200).json({ success: true, data: response, message: "Awareness Category Updated", });
@@ -37,6 +66,15 @@ const getAllCategory = async (req, res) => {
 
 const deleteCategory = async (req, res) => {
   try {
+
+    const checkAwarenessCategory = await AwarenessCategory.findById(req.params.id)
+    if (!checkAwarenessCategory) {
+      return res.status(404).json({ success: false, message: "Category not found" });
+    }
+    if (checkAwarenessCategory.image) {
+      await deleteImgFromFolder(checkAwarenessCategory.image, "awarenessCategory");
+    }
+
     const response = await AwarenessCategory.findByIdAndDelete(req.params.id);
     if (response) {
       return res.status(200).json({ success: true, message: "Awareness Category deleted" });
@@ -48,13 +86,20 @@ const deleteCategory = async (req, res) => {
 };
 
 const CreateAwareNess = async (req, res) => {
+  // console.log(" ========================== CreateAwareNess ==========================")
+  const image = req.files?.image
   try {
+    if (image) {
+      const imagePath = await UploadImage(image, "awareness");
+      req.body.image = imagePath;
+    }
     const response = await Awareness.create(req.body);
     if (response) {
-      res.status(201).json({ success: true, data: response, message: "Awareness Uploaded" });
+      return res.status(201).json({ success: true, data: response, message: "Awareness Uploaded" });
     } else {
-      res.status(400).json({ success: false, message: "Awareness not Uploaded" });
+      return res.status(400).json({ success: false, message: "Awareness not Uploaded" });
     }
+    // res.status(400).json({ success: false, message: "Awareness not Uploaded" });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -99,7 +144,23 @@ const getAwarenessByCategory = async (req, res) => {
 };
 
 const UpdateAwareness = async (req, res) => {
+  // console.log(" ========================== UpdateAwareNess ==========================")
+  // console.log("req.body: ", req.body);
+  // console.log("req.files: ", req.files);
+
   try {
+    const checkAwareness = await Awareness.findById(req.params.id);
+
+    if (!checkAwareness) {
+      return res.status(404).json({ success: false, message: "Awareness not found" });
+    }
+    if (req.files?.image) {
+      const imagePath = await UploadImage(req.files.image, "awareness");
+      req.body.image = imagePath;
+      if (checkAwareness.image) {
+        await deleteImgFromFolder(checkAwareness.image, "awareness");
+      }
+    }
     const response = await Awareness.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true, });
     if (response) {
       return res.status(200).json({ success: true, data: response, message: "Awareness  Updated", });
@@ -111,11 +172,20 @@ const UpdateAwareness = async (req, res) => {
 };
 const deleteAwareness = async (req, res) => {
   try {
+    const checkAwareness = await Awareness.findById(req.params.id);
+
+    if (!checkAwareness) {
+      return res.status(404).json({ success: false, message: "Awareness not found" });
+    }
+    if (checkAwareness.image) {
+      await deleteImgFromFolder(checkAwareness.image, "awareness");
+    }
+
     const response = await Awareness.findByIdAndDelete(req.params.id);
     if (response) {
-      res.status(200).json({ success: true, message: "Awareness  deleted" });
+      return res.status(200).json({ success: true, message: "Awareness  deleted" });
     } else {
-      res.status(404).json({ success: false, message: "Awareness not found" });
+      return res.status(404).json({ success: false, message: "Awareness not found" });
     }
   } catch (error) {
     res.status(500).json({ message: error.message });
